fix(node-ui): correct edge_update id, dedup output helpers, handle agent updates
Node UI / state / event-streaming consistency fixes:

- edge_update: kernel stamped the run id into `workflow_id`; the unified
  websocket runner only backfills missing fields, so the wrong value leaked
  through. Emit `job_id` instead and let `workflow_id` be backfilled from the
  active run. EdgeUpdate now carries both (optional) ids.

- Output value extraction: PreviewNode and OutputNode each carried an identical
  copy of getOutputFromResult/getCopySource, and ChainNodeCard a third, simpler
  copy that dropped array and `.value` handling (so chain cards mis-rendered
  array/.value outputs). Extracted to a shared node/outputResult.ts; all three
  now behave identically.

- Agent step_result / todo_update: step_result was in the workflow MsgpackData
  union with no handler, and todo_update was missing from the union entirely, so
  both were silently dropped in the workflow path (only the chat path handled
  them). Added TodoUpdate to the union and recorded both on the trace timeline
  alongside tool_call/tool_result.
